requirements.txt
Flask
google-generativeai
pypdf
python-dotenv
Werkzeug
